+++ web/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2023-03-25
+## [1.0] - 2023-03-25
 
 ### Fixed
 
@@ -66,3 +66,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - cache data from `sessionStorage`
+
+## [1.1] - 2023-04-08
+
+### Added
+
+- lazyload: disable skip image if current page is chapter
+- lazyload: get image dimensions before image has fully loaded
+- connection-notif: show/hide click on chapter page